Finished 1.8.9 code, fixes issue #22, fixes issue #31

+++ library/src/main/java/org/apache/pdfbox/pdmodel/documentinterchange/taggedpdf/PDLayoutAttributeObject.java
@@ -1502,7 +1502,7 @@ public class PDLayoutAttributeObject extends PDStandardAttributeObject
             }
             else
             {
-                sb.append(String.valueOf((Float) borderThickness));
+                sb.append(String.valueOf(borderThickness));
             }
         }
         if (this.isSpecified(PADDING))
@@ -1515,7 +1515,7 @@ public class PDLayoutAttributeObject extends PDStandardAttributeObject
             }
             else
             {
-                sb.append(String.valueOf((Float) padding));
+                sb.append(String.valueOf(padding));
             }
         }
         if (this.isSpecified(COLOR))
@@ -1561,7 +1561,7 @@ public class PDLayoutAttributeObject extends PDStandardAttributeObject
             sb.append(", Width=");
             if (width instanceof Float)
             {
-                sb.append(String.valueOf((Float) width));
+                sb.append(String.valueOf(width));
             }
             else
             {
@@ -1574,7 +1574,7 @@ public class PDLayoutAttributeObject extends PDStandardAttributeObject
             sb.append(", Height=");
             if (height instanceof Float)
             {
-                sb.append(String.valueOf((Float) height));
+                sb.append(String.valueOf(height));
             }
             else
             {
@@ -1612,7 +1612,7 @@ public class PDLayoutAttributeObject extends PDStandardAttributeObject
             }
             else
             {
-                sb.append(String.valueOf((Float) tPadding));
+                sb.append(String.valueOf(tPadding));
             }
         }
         if (this.isSpecified(BASELINE_SHIFT))
@@ -1626,7 +1626,7 @@ public class PDLayoutAttributeObject extends PDStandardAttributeObject
             sb.append(", LineHeight=");
             if (lineHeight instanceof Float)
             {
-                sb.append(String.valueOf((Float) lineHeight));
+                sb.append(String.valueOf(lineHeight));
             }
             else
             {
@@ -1676,7 +1676,7 @@ public class PDLayoutAttributeObject extends PDStandardAttributeObject
             }
             else
             {
-                sb.append(String.valueOf((Float) columnGap));
+                sb.append(String.valueOf(columnGap));
             }
         }
         if (this.isSpecified(COLUMN_WIDTHS))
@@ -1689,7 +1689,7 @@ public class PDLayoutAttributeObject extends PDStandardAttributeObject
             }
             else
             {
-                sb.append(String.valueOf((Float) columnWidth));
+                sb.append(String.valueOf(columnWidth));
             }
         }
         return sb.toString();
